Move storage to its own data class, move authorisedRepository to its own class
